portfolio_app/requirements.txt
streamlit>=1.36.0
plotly>=5.15.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
cvxpy>=1.4.0
yfinance>=0.2.30
scikit-learn>=1.3.0
statsmodels>=0.14.0
PyYAML>=6.0
